+++ src/type/type.ts
@@ -20,8 +20,3 @@ export const DEBATE_TYPE_LABELS: Record<DebateType, string> = {
   CLOSING: '최종발언',
   TIME_OUT: '작전 시간',
 };
-
-export interface LinkButtonProps {
-  url: string;
-  title: string;
-}
